Split up the save/load handling functions from PrefabIntegrationManager to improve maintainability.
