+++ Code/Framework/AzToolsFramework/AzToolsFramework/UI/Prefab/PrefabIntegrationManager.h
@@ -9,22 +9,17 @@
 #pragma once
 
 #include <AzCore/Memory/SystemAllocator.h>
-#include <AzCore/UserSettings/UserSettings.h>
 
 #include <AzToolsFramework/API/ToolsApplicationAPI.h>
-#include <AzToolsFramework/AssetBrowser/AssetBrowserSourceDropBus.h>
 #include <AzToolsFramework/Editor/EditorContextMenuBus.h>
 #include <AzToolsFramework/Entity/EditorEntityContextBus.h>
-#include <AzToolsFramework/Prefab/PrefabPublicInterface.h>
-#include <AzToolsFramework/Prefab/PrefabSystemComponentInterface.h>
 #include <AzToolsFramework/UI/Prefab/LevelRootUiHandler.h>
 #include <AzToolsFramework/UI/Prefab/PrefabIntegrationBus.h>
 #include <AzToolsFramework/UI/Prefab/PrefabIntegrationInterface.h>
+#include <AzToolsFramework/UI/Prefab/PrefabSaveLoadHandler.h>
 #include <AzToolsFramework/UI/Prefab/PrefabUiHandler.h>
 #include <AzToolsFramework/UI/Prefab/Procedural/ProceduralPrefabReadOnlyHandler.h>
 #include <AzToolsFramework/UI/Prefab/Procedural/ProceduralPrefabUiHandler.h>
-
-#include <AzQtComponents/Components/Widgets/Card.h>
 
 namespace AzToolsFramework
 {
@@ -35,30 +30,13 @@ namespace AzToolsFramework
     {
         class PrefabFocusPublicInterface;
         class PrefabLoaderInterface;
-
-        //! Structure for saving/retrieving user settings related to prefab workflows.
-        class PrefabUserSettings
-            : public AZ::UserSettings
-        {
-        public:
-            AZ_CLASS_ALLOCATOR(PrefabUserSettings, AZ::SystemAllocator, 0);
-            AZ_RTTI(PrefabUserSettings, "{E17A6128-E2C3-4501-B1AD-B8BB0D315602}", AZ::UserSettings);
-
-            AZStd::string m_saveLocation;
-            bool m_autoNumber = false; //!< Should the name of the prefab file be automatically numbered. e.g PrefabName_001.prefab vs PrefabName.prefab.
-
-            PrefabUserSettings() = default;
-
-            static void Reflect(AZ::ReflectContext* context);
-        };
+        class PrefabPublicInterface;
 
         class PrefabIntegrationManager final
             : public EditorContextMenuBus::Handler
             , public EditorEventsBus::Handler
-            , public AssetBrowser::AssetBrowserSourceDropBus::Handler
             , public PrefabInstanceContainerNotificationBus::Handler
             , public PrefabIntegrationInterface
-            , public QObject
             , private EditorEntityContextNotificationBus::Handler
         {
         public:
@@ -77,9 +55,6 @@ namespace AzToolsFramework
             // EditorEventsBus overrides ...
             void OnEscape() override;
 
-            // EntityOutlinerSourceDropHandlingBus overrides ...
-            void HandleSourceFileType(AZStd::string_view sourceFilePath, AZ::EntityId parentId, AZ::Vector3 position) const override;
-
             // EditorEntityContextNotificationBus overrides ...
             void OnStartPlayInEditorBegin() override;
             void OnStopPlayInEditor() override;
@@ -94,6 +69,9 @@ namespace AzToolsFramework
             void ExecuteSavePrefabDialog(TemplateId templateId, bool useSaveAllPrefabsPreference) override;
 
         private:
+            // Handles the UI for prefab save operations.
+            PrefabSaveHandler m_prefabSaveHandler;
+
             // Used to handle the UI for the level root.
             LevelRootUiHandler m_levelRootUiHandler;
 
@@ -120,27 +98,6 @@ namespace AzToolsFramework
             void InitializeShortcuts();
             void UninitializeShortcuts();
 
-            // Prompt and resolve dialogs
-            static bool QueryUserForPrefabSaveLocation(
-                const AZStd::string& suggestedName, const char* initialTargetDirectory, AZ::u32 prefabUserSettingsId, QWidget* activeWindow,
-                AZStd::string& outPrefabName, AZStd::string& outPrefabFilePath);
-            static bool QueryUserForPrefabFilePath(AZStd::string& outPrefabFilePath);
-            static bool QueryUserForProceduralPrefabAsset(AZStd::string& outPrefabAssetPath);
-            static void WarnUserOfError(AZStd::string_view title, AZStd::string_view message);
-
-            // Path and filename generation
-            static void GenerateSuggestedFilenameFromEntities(const EntityIdList& entities, AZStd::string& outName);
-            static bool AppendEntityToSuggestedFilename(AZStd::string& filename, AZ::EntityId entityId);
-
-            enum class PrefabSaveResult
-            {
-                Continue,
-                Retry,
-                Cancel
-            };
-            static PrefabSaveResult IsPrefabPathValidForAssets(QWidget* activeWindow, QString prefabPath, AZStd::string& retrySavePath);
-            static void GenerateSuggestedPrefabPath(const AZStd::string& prefabName, const AZStd::string& targetDirectory, AZStd::string& suggestedFullPath);
-
             // Reference detection
             static void GatherAllReferencedEntitiesAndCompare(const EntityIdSet& entities, EntityIdSet& entitiesAndReferencedEntities,
                 bool& hasExternalReferences);
@@ -148,20 +105,10 @@ namespace AzToolsFramework
             static bool QueryAndPruneMissingExternalReferences(EntityIdSet& entities, EntityIdSet& selectedAndReferencedEntities,
                 bool& useReferencedEntities, bool defaultMoveExternalRefs = false);
 
-            // Settings management
-            static void SetPrefabSaveLocation(const AZStd::string& path, AZ::u32 settingsId);
-            static bool GetPrefabSaveLocation(AZStd::string& path, AZ::u32 settingsId);
-
             static AZ::u32 GetSliceFlags(const AZ::Edit::ElementData* editData, const AZ::Edit::ClassData* classData);
-
-            AZStd::shared_ptr<QDialog> ConstructClosePrefabDialog(TemplateId templateId);
-            AZStd::unique_ptr<AzQtComponents::Card> ConstructUnsavedPrefabsCard(TemplateId templateId);
-            AZStd::unique_ptr<QDialog> ConstructSavePrefabDialog(TemplateId templateId, bool useSaveAllPrefabsPreference);
-            void SavePrefabsInDialog(QDialog* unsavedPrefabsDialog);
 
             AZStd::vector<AZStd::unique_ptr<QAction>> m_actions;
 
-            static const AZStd::string s_prefabFileExtension;
             static AzFramework::EntityContextId s_editorEntityContextId;
 
             static ContainerEntityInterface* s_containerEntityInterface;
@@ -169,7 +116,6 @@ namespace AzToolsFramework
             static PrefabFocusPublicInterface* s_prefabFocusPublicInterface;
             static PrefabLoaderInterface* s_prefabLoaderInterface;
             static PrefabPublicInterface* s_prefabPublicInterface;
-            static PrefabSystemComponentInterface* s_prefabSystemComponentInterface;
 
             ReadOnlyEntityPublicInterface* m_readOnlyEntityPublicInterface = nullptr;
         };
